Transition from ConfigParser to our own subclass
Subclass RawConfigParser ourselves and use that, to allow for
better handling of defaults, as well as per-platform config files.
It also allows us to override getboolean() and always return a sane
value, avoiding a bunch of try/except blocks.

Add some new config values, including whether or not we should
syslog when lockers are invalid or not found.

Add a sample config file